+++ src/filter/soloud_biquadresonantfilter.cpp
@@ -100,7 +100,7 @@ namespace SoLoud
 		calcBQRParams();
 	}
 
-	void BiquadResonantFilterInstance::filterChannel(float *aBuffer, int aSamples, float aSamplerate, float aTime, int aChannel, int aChannels)
+	void BiquadResonantFilterInstance::filterChannel(float *aBuffer, int aSamples, float aSamplerate, double aTime, int aChannel, int aChannels)
 	{
 		if (!mActive)
 			return;
